requirements.txt
appdirs==1.4.3
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
blis==0.4.1
cached-property==1.5.1
cachetools==4.0.0
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
Flask==1.1.1
future==0.18.2
fuzzywuzzy==0.3.0
google-ads==4.1.1
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleads==23.0.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
httplib2==0.17.0
idna==2.8
importlib-metadata==1.4.0
ipython==7.13.0
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.10.3
linkedin-api==1.1.0
lxml==4.4.2
MarkupSafe==1.1.1
more-itertools==8.1.0
murmurhash==1.0.2
nltk==3.4.5
numpy==1.18.1
oauth2client==4.1.3
oauthlib==3.1.0
parso==0.7.0
pdf2image==1.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plac==1.1.3
preshed==3.0.2
prompt-toolkit==3.0.5
protobuf==3.11.3
psycopg2==2.8.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pytesseract==0.3.1
python-docx==0.8.10
python-linkedin-v2==0.9.4
pytz==2019.3
PyYAML==5.3
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.14.0
soupsieve==1.9.5
spacy==2.2.3
SQLAlchemy==1.3.16
srsly==1.0.1
thinc==7.3.1
tqdm==4.42.0
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.8
wasabi==0.6.0
wcwidth==0.1.9
Werkzeug==1.0.0
xmltodict==0.12.0
zeep==3.4.0
zipp==2.0.1
